backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "channels>=4.3.1",
    "daphne>=4.2.1",
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.1",
    "django-oauth-toolkit>=3.0.1",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "markdown>=3.9",
    "psycopg2-binary>=2.9.10",
    "social-auth-app-django>=5.5.1",
]
